+++ app/javascript/vue/tasks/otu/browse/store/mutations/setObservationsDepictions.js
@@ -1,0 +1,3 @@
+export default (state, value) => {
+  state.observationsDepictions = value
+}